close #5

+++ src/formula/PilotoDecorador.java
@@ -50,19 +50,19 @@ public abstract class PilotoDecorador extends PilotoBase implements Serializable
         piloto.setEdad(edad);
     }
 
-    public Double getAltura() {
+    public Integer getAltura() {
         return piloto.getAltura();
     }
 
-    public void setAltura(Double altura) {
+    public void setAltura(Integer altura) {
         piloto.setAltura(altura);
     }
 
-    public Double getPeso() {
+    public Integer getPeso() {
         return piloto.getPeso();
     }
 
-    public void setPeso(Double peso) {
+    public void setPeso(Integer peso) {
         piloto.setPeso(peso);
     }
